% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simdat.R
\docType{data}
\name{simdat_pic}
\alias{simdat_pic}
\title{Simulated clustered partly interval-censored data}
\format{
A data frame with 6 variables:
\describe{
\item{L}{Left endpoint of observation interval}
\item{R}{Right endpoint of observation interval}
\item{delta}{Exact observation indicator (1 = exact, 0 = interval)}
\item{id}{Cluster identifier}
\item{x1}{First covariate}
\item{x2}{Second covariate}
}
}
\source{
Simulated data
}
\usage{
simdat_pic
}
\description{
Example dataset for penalized AFT model fitting with clustered partly interval-censored survival data.
}
\keyword{datasets}
