% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simdat.R
\docType{data}
\name{simdat_rc}
\alias{simdat_rc}
\title{Simulated right-censored data}
\format{
A data frame with 100 rows and 12 variables:
\describe{
\item{y}{Observed survival or censoring time}
\item{d}{Censoring indicator (1 = event observed, 0 = censored)}
\item{X1}{First covariate}
\item{X2}{Second covariate}
\item{X3}{Third covariate}
\item{X4}{Fourth covariate}
\item{X5}{Fifth covariate}
\item{X6}{Sixth covariate}
\item{X7}{Seventh covariate}
\item{X8}{Eighth covariate}
\item{X9}{Ninth covariate}
\item{X10}{Tenth covariate}
}
}
\source{
Simulated data
}
\usage{
simdat_rc
}
\description{
Example dataset for penalized AFT model fitting with right-censored survival data.
}
\keyword{datasets}
